Undoing SecurityUtils because it's too much churn
