+++ src/main/java/org/sagebionetworks/bridge/hibernate/HibernateAccountDao.java
@@ -21,7 +21,6 @@ import com.google.common.collect.ImmutableMap;
 
 import org.sagebionetworks.bridge.BridgeUtils;
 import org.sagebionetworks.bridge.RequestContext;
-import org.sagebionetworks.bridge.SecurityUtils;
 import org.sagebionetworks.bridge.dao.AccountDao;
 import org.sagebionetworks.bridge.time.DateUtils;
 import org.sagebionetworks.bridge.models.AccountSummarySearch;
@@ -234,15 +233,15 @@ public class HibernateAccountDao implements AccountDao {
         // Hibernate will not load the collection of substudies once you use the constructor form of HQL 
         // to limit the data you retrieve from a table. May need to manually construct the objects to 
         // avoid this 1+N query.
-        SecurityUtils.SubstudyAssociations assoc = null;
+        BridgeUtils.SubstudyAssociations assoc = null;
         if (hibernateAccount.getId() != null) {
             List<HibernateAccountSubstudy> accountSubstudies = hibernateHelper.queryGet(
                     "FROM HibernateAccountSubstudy WHERE accountId=:accountId",
                     ImmutableMap.of("accountId", hibernateAccount.getId()), null, null, HibernateAccountSubstudy.class);
             
-            assoc = SecurityUtils.substudyAssociationsVisibleToCaller(accountSubstudies);
+            assoc = BridgeUtils.substudyAssociationsVisibleToCaller(accountSubstudies);
         } else {
-            assoc = SecurityUtils.substudyAssociationsVisibleToCaller(null);
+            assoc = BridgeUtils.substudyAssociationsVisibleToCaller(null);
         }
         
         return new AccountSummary(hibernateAccount.getFirstName(), hibernateAccount.getLastName(),
