+++ src/main/java/org/sagebionetworks/bridge/models/accounts/UserSession.java
@@ -5,8 +5,8 @@ import static com.google.common.base.Preconditions.checkNotNull;
 import java.util.Map;
 import java.util.Set;
 
+import org.sagebionetworks.bridge.BridgeUtils;
 import org.sagebionetworks.bridge.Roles;
-import org.sagebionetworks.bridge.SecurityUtils;
 import org.sagebionetworks.bridge.config.Environment;
 import org.sagebionetworks.bridge.models.studies.StudyIdentifier;
 import org.sagebionetworks.bridge.models.subpopulations.SubpopulationGuid;
@@ -97,10 +97,10 @@ public class UserSession {
         return ConsentStatus.isConsentCurrent(consentStatuses);
     }
     public boolean isInRole(Roles role) {
-        return SecurityUtils.isInRole(participant.getRoles(), role);
+        return BridgeUtils.isInRole(participant.getRoles(), role);
     }
     public boolean isInRole(Set<Roles> roleSet) {
-        return SecurityUtils.isInRole(participant.getRoles(), roleSet);
+        return BridgeUtils.isInRole(participant.getRoles(), roleSet);
     }
     // These are accessed so frequently it is worth having convenience accessors
     @JsonIgnore
