+++ src/main/java/org/sagebionetworks/bridge/services/AccountService.java
@@ -26,7 +26,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import org.sagebionetworks.bridge.BridgeUtils;
-import org.sagebionetworks.bridge.SecurityUtils;
 import org.sagebionetworks.bridge.dao.AccountDao;
 import org.sagebionetworks.bridge.dao.AccountSecretDao;
 import org.sagebionetworks.bridge.exceptions.AccountDisabledException;
@@ -267,7 +266,7 @@ public class AccountService {
         Optional<Account> optional = accountDao.getAccount(accountId);
         if (optional.isPresent()) {
             // filtering based on the substudy associations of the caller.
-            return SecurityUtils.filterForSubstudy(optional.get());
+            return BridgeUtils.filterForSubstudy(optional.get());
         }
         return null;
     }
