+++ src/main/java/org/sagebionetworks/bridge/services/AssessmentResourceService.java
@@ -5,8 +5,8 @@ import static com.google.common.base.Preconditions.checkNotNull;
 import static java.util.stream.Collectors.toList;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
 import static org.jsoup.safety.Whitelist.simpleText;
+import static org.sagebionetworks.bridge.BridgeUtils.checkOwnership;
 import static org.sagebionetworks.bridge.BridgeUtils.sanitizeHTML;
-import static org.sagebionetworks.bridge.SecurityUtils.checkOwnership;
 import static org.sagebionetworks.bridge.models.ResourceList.CATEGORIES;
 import static org.sagebionetworks.bridge.models.ResourceList.INCLUDE_DELETED;
 import static org.sagebionetworks.bridge.models.ResourceList.MAX_REVISION;
